+++ AetherEngine/RenderingEngine/CMakeLists.txt
@@ -17,12 +17,15 @@ set(SRC
 
     Window/src/WindowGLFW.cpp
     
-    Renderer/DirectX/src/RendererDX11.cpp
-    Renderer/DirectX/src/RendererDX12.cpp
-    Renderer/OpenGL/src/RendererOpenGL.cpp
-     
     Graphics/src/GraphicsContext.cpp
     Graphics/src/Shader.cpp
+
+    Renderer/DirectX/src/RendererDX11.cpp
+    Renderer/DirectX/src/RendererDX12.cpp
+    Renderer/DirectX/src/ShaderDX12.cpp
+
+    Renderer/OpenGL/src/RendererOpenGL.cpp
+    Renderer/OpenGL/src/ShaderOpenGL.cpp
 
     Layers/src/ImGuiLayer.cpp
 )
@@ -52,15 +55,24 @@ set(INCLUDE
 
     Window/include/WindowGLFW.h
 
+    #
     # Rendering
-    Renderer/DirectX/include/D3DUtils.h
-    Renderer/DirectX/include/RendererDX11.h
-    Renderer/DirectX/include/RendererDX12.h
-    Renderer/OpenGL/include/RendererOpenGL.h
+    #
 
+    
     # Graphics
     Graphics/include/GraphicsContext.h
     Graphics/include/Shader.h
+
+    # DirectX
+    Renderer/DirectX/include/D3DUtils.h
+    Renderer/DirectX/include/RendererDX11.h
+    Renderer/DirectX/include/RendererDX12.h
+    Renderer/DirectX/include/ShaderDX12.h
+
+    # OpenGL
+    Renderer/OpenGL/include/RendererOpenGL.h
+    Renderer/OpenGL/include/ShaderOpenGL.h
 
     # Events
     Events/include/KeyEvent.h
