+++ AetherEngine/RenderingEngine/Core/include/CoreApp.h
@@ -46,7 +46,7 @@ namespace Aether
 		ImGuiLayer* m_ImGuiLayer = nullptr;
 
 		LayerStack m_LayerStack;
-		eRenderAPI m_CurrentRenderAPI = eRenderAPI::kOpenGL;
+		eRenderAPI m_CurrentRenderAPI = eRenderAPI::kDX12;
 
 		// Quick and dirty Singleton-esque implementation
 		static Application* s_Instance;
